fix: フレームの消火
