--- src/features/profile/badges/components/badge/BadgeFlame.tsx
@@ -1,403 +1,0 @@
-import { FC } from "react";
-
-type Props = {
-  className: string;
-};
-
-export const BadgeFlame: FC<Props> = ({ className }) => {
-  return (
-    <svg
-      version="1.1"
-      id="flame"
-      xmlns="http://www.w3.org/2000/svg"
-      xmlnsXlink="http://www.w3.org/1999/xlink"
-      x="0px"
-      y="0px"
-      width="100%"
-      viewBox="0 0 500 500"
-      enableBackground="new 0 0 500 500"
-      xmlSpace="preserve"
-      className={className}
-    >
-      <path
-        fill="#F7CE01"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M157.286118,72.270378 
-        C173.516327,63.263542 190.491089,57.060749 208.151581,53.167286 
-        C229.759186,48.403625 251.701157,47.991753 273.774048,49.815792 
-        C290.873474,51.228848 307.304474,55.559956 323.186981,61.336189 
-        C335.297791,65.740692 346.874908,72.060837 357.937073,78.769432 
-        C368.189056,84.986687 378.133087,92.093834 386.973450,100.166290 
-        C399.957916,112.022827 412.046539,124.994774 420.817810,140.431839 
-        C427.691437,152.529053 434.348877,164.834488 439.911774,177.567551 
-        C445.724182,190.871613 449.195862,205.053192 450.957184,219.527344 
-        C452.182922,229.600327 453.987701,239.740250 453.811890,249.820236 
-        C453.313965,278.364716 448.299438,306.181549 435.842377,332.098389 
-        C429.991577,344.270935 422.817535,355.825104 415.908417,367.460785 
-        C413.799347,371.012665 413.592590,373.247681 416.720642,376.363159 
-        C427.898987,387.496796 438.767975,398.941071 450.364685,410.897400 
-        C444.378998,411.661530 439.174530,412.550293 433.932800,412.948212 
-        C423.099060,413.770691 412.238403,414.235504 401.401855,415.025635 
-        C389.117767,415.921265 376.835785,416.892334 364.583618,418.133820 
-        C362.504547,418.344452 360.513000,419.969788 358.594513,421.129333 
-        C339.041077,432.947296 318.094238,441.415924 295.872711,446.529388 
-        C287.967468,448.348541 279.796143,449.261292 271.692078,449.901794 
-        C262.140930,450.656677 252.509369,451.242706 242.950775,450.884674 
-        C220.745102,450.052948 199.315384,445.191315 178.551971,437.220093 
-        C160.245361,430.192047 143.525284,420.387726 128.077850,408.467224 
-        C113.694069,397.367584 101.157005,384.307281 90.353577,369.705292 
-        C76.573471,351.080017 65.903740,330.723877 59.918159,308.320984 
-        C56.378944,295.074402 54.042160,281.436249 52.184326,267.834900 
-        C50.892151,258.374786 50.596264,248.635178 51.149662,239.095535 
-        C52.404667,217.461349 55.988621,196.135162 65.069313,176.264130 
-        C72.164627,160.737625 80.216789,145.648376 88.107613,129.836166 
-        C76.648560,117.558258 64.276131,104.301682 50.905609,89.975700 
-        C58.875954,89.538963 65.686455,89.280914 72.477432,88.765625 
-        C81.757019,88.061493 91.021973,87.164207 100.292503,86.341881 
-        C101.271278,86.255058 102.242836,86.048317 103.222069,86.011177 
-        C112.479996,85.659988 121.740807,85.379257 130.996841,84.985451 
-        C138.357803,84.672264 143.617645,79.531517 149.818268,76.560852 
-        C152.287643,75.377792 154.597946,73.862694 157.286118,72.270378 
-      M360.753326,316.676697 
-        C361.196045,316.776611 361.638733,316.876556 362.012848,316.995789 
-        C362.012848,316.995789 362.044708,316.932037 361.559448,317.536560 
-        C366.976257,323.311462 372.393066,329.086365 378.031616,335.097656 
-        C379.215912,333.357300 380.099365,332.175354 380.862732,330.920319 
-        C401.881561,296.362793 408.814453,259.208740 400.986511,219.488480 
-        C388.054138,153.867371 332.686707,103.065140 264.206116,97.246574 
-        C230.667465,94.396904 199.202850,101.761490 170.345840,119.566704 
-        C167.005951,121.627464 165.489288,123.986473 165.030243,127.760429 
-        C163.665237,138.982697 161.985306,150.166382 160.507004,161.375534 
-        C160.352097,162.550034 160.719940,163.793457 160.986374,165.852829 
-        C159.287949,177.519653 157.589508,189.186478 155.619873,202.716171 
-        C149.085907,195.844574 143.539993,190.012115 138.064575,183.466309 
-        C132.976929,177.896500 127.889267,172.326691 122.663727,166.605927 
-        C121.824883,167.749680 121.391296,168.257095 121.048729,168.819931 
-        C101.276909,201.305206 93.990128,236.315262 99.952454,273.948517 
-        C106.103142,312.770813 124.258728,345.268036 155.260269,369.026398 
-        C207.510910,409.069153 263.992950,413.452209 323.279449,384.921936 
-        C332.254974,380.602661 338.183044,375.300232 337.834167,364.493378 
-        C337.625031,358.013611 338.669647,351.493408 339.029694,344.160492 
-        C339.337891,343.422211 339.835785,342.706146 339.925415,341.942230 
-        C340.985718,332.902039 341.921021,323.846832 343.028107,314.812592 
-        C343.558411,310.485443 344.383484,306.194397 345.167053,301.313995 
-        C350.383362,306.407104 354.773956,310.694031 359.283173,315.719238 
-        C359.875183,315.794281 360.467163,315.869324 360.999420,315.988647 
-        C360.999420,315.988647 361.023956,315.918457 360.753326,316.676697 
-      M331.780304,95.687363 
-        C342.159668,102.868599 353.023468,109.451630 362.803833,117.372505 
-        C377.158875,128.998337 388.052460,143.784073 397.879028,159.370346 
-        C401.718750,165.460648 407.261108,166.761993 412.309784,163.440826 
-        C417.377350,160.107239 418.296478,154.604385 414.600830,148.556137 
-        C395.693665,117.612885 370.801208,92.969185 338.297577,76.412148 
-        C323.102875,68.672112 307.102539,63.131424 290.477264,59.398651 
-        C284.858215,58.137039 279.691620,61.053478 278.195282,66.139580 
-        C276.706390,71.200439 279.101257,75.925270 284.450165,78.170921 
-        C286.274963,78.937027 288.281342,79.259300 290.142700,79.951309 
-        C303.823547,85.037468 317.488678,90.165962 331.780304,95.687363 
-      M100.638000,368.839569 
-        C107.407326,375.461517 113.974586,382.312897 121.071205,388.563263 
-        C123.153931,390.397614 126.765190,391.336060 129.629028,391.242950 
-        C133.705093,391.110504 136.662506,388.319946 137.766647,384.222076 
-        C139.126160,379.176422 136.710892,375.577454 132.903351,372.592377 
-        C117.999825,360.907928 107.236870,346.221619 100.636574,328.434326 
-        C98.135201,321.693359 92.958626,318.919922 87.630302,320.958221 
-        C81.457748,323.319489 78.795601,329.033264 82.134781,335.691681 
-        C87.697586,346.784180 94.120689,357.445221 100.638000,368.839569 
-      M428.938232,186.311630 
-        C431.944855,180.461761 431.447693,175.793213 427.446198,172.300903 
-        C424.096100,169.377090 418.531738,169.041290 414.902557,171.543930 
-        C410.277100,174.733597 409.142578,182.204514 412.611084,186.633514 
-        C416.280518,191.319092 423.089508,191.441910 428.938232,186.311630 
-      M85.004288,297.575378 
-        C80.608101,294.235229 74.696617,293.608826 71.201172,296.341278 
-        C67.579636,299.172302 66.221504,302.938538 67.564117,307.427948 
-        C69.246780,313.054352 75.372406,316.219696 80.557129,314.246002 
-        C87.271980,311.689758 89.186470,305.516693 85.004288,297.575378 
-      z"
-      />
-      <path
-        fill="#F6F5F4"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M361.802063,317.234314 
-        C363.365234,315.544861 365.125153,314.375397 365.937225,312.736328 
-        C372.811157,298.861023 378.603912,284.610687 379.985657,268.960968 
-        C380.647217,261.467896 382.082825,253.949661 381.829163,246.488342 
-        C381.260864,229.772293 377.768646,213.657211 371.138702,198.096909 
-        C364.618225,182.793610 355.461914,169.307053 343.924500,157.741974 
-        C326.770020,140.546356 306.254547,128.638321 282.346558,122.981377 
-        C269.690460,119.986778 256.901154,118.152130 243.943466,119.103752 
-        C228.616470,120.229370 213.446045,122.801247 199.619781,129.940247 
-        C189.388367,135.223099 179.448898,141.117752 169.662064,147.195099 
-        C165.000580,150.089752 161.784775,154.450836 161.964798,160.525864 
-        C162.000626,161.735214 161.335632,162.965347 160.919525,164.595398 
-        C160.719940,163.793457 160.352097,162.550034 160.507004,161.375534 
-        C161.985306,150.166382 163.665237,138.982697 165.030243,127.760429 
-        C165.489288,123.986473 167.005951,121.627464 170.345840,119.566704 
-        C199.202850,101.761490 230.667465,94.396904 264.206116,97.246574 
-        C332.686707,103.065140 388.054138,153.867371 400.986511,219.488480 
-        C408.814453,259.208740 401.881561,296.362793 380.862732,330.920319 
-        C380.099365,332.175354 379.215912,333.357300 378.031616,335.097656 
-        C372.393066,329.086365 366.976257,323.311462 361.802063,317.234314 
-      z"
-      />
-      <path
-        fill="#F5F5F4"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M280.985046,377.154175 
-        C281.703827,377.096313 282.422607,377.038422 283.726990,376.890503 
-        C284.613037,376.664185 284.837189,376.448639 284.985138,376.153809 
-        C299.801208,372.593231 313.076843,365.711761 325.399963,356.969452 
-        C330.191132,353.570465 334.488617,349.475616 339.080872,345.344543 
-        C338.669647,351.493408 337.625031,358.013611 337.834167,364.493378 
-        C338.183044,375.300232 332.254974,380.602661 323.279449,384.921936 
-        C263.992950,413.452209 207.510910,409.069153 155.260269,369.026398 
-        C124.258728,345.268036 106.103142,312.770813 99.952454,273.948517 
-        C93.990128,236.315262 101.276909,201.305206 121.048729,168.819931 
-        C121.391296,168.257095 121.824883,167.749680 122.663727,166.605927 
-        C127.889267,172.326691 132.976929,177.896500 138.029327,183.822968 
-        C127.960106,200.782745 122.121040,218.707062 120.593582,238.635056 
-        C120.161827,240.450684 120.142166,241.750381 120.122505,243.050079 
-        C120.089645,247.375992 120.056786,251.701920 119.810623,256.673340 
-        C120.013550,258.210510 120.429779,259.102142 120.846008,259.993805 
-        C120.903778,262.326477 120.961555,264.659119 121.109253,267.651672 
-        C121.335281,268.612213 121.550728,268.836700 121.845505,268.985016 
-        C121.902687,270.340515 121.959877,271.696045 122.108047,273.682556 
-        C122.335533,274.613281 122.550911,274.837067 122.845154,274.984955 
-        C122.903473,276.018036 122.961792,277.051147 123.109711,278.697571 
-        C123.335274,279.612030 123.550880,279.836761 123.846130,279.985046 
-        C123.903893,280.703857 123.961655,281.422638 124.109482,282.727051 
-        C124.335777,283.613037 124.551300,283.837189 124.846123,283.985077 
-        C128.865906,299.606903 135.675949,313.965515 145.392441,326.774902 
-        C165.749115,353.611328 191.991837,371.479614 225.429001,377.972504 
-        C227.654877,378.404724 229.915466,378.658173 232.727966,379.266541 
-        C233.538147,379.750275 233.816666,379.889343 234.131577,379.955566 
-        C236.796921,379.968506 239.462265,379.981415 242.661224,380.434204 
-        C245.480194,380.874512 247.765564,380.875031 250.050934,380.875488 
-        C250.472519,380.908783 250.894119,380.942108 251.883316,381.220062 
-        C254.631897,381.027832 256.812866,380.590973 258.993835,380.154083 
-        C261.680603,380.094025 264.367371,380.033966 267.655640,380.160522 
-        C268.836060,379.949219 269.414978,379.551331 269.993866,379.153412 
-        C271.347778,379.096619 272.701721,379.039825 274.685822,378.892029 
-        C275.615143,378.663879 275.838196,378.448059 275.985168,378.153595 
-        C277.019226,378.096100 278.053284,378.038635 279.700256,377.890930 
-        C280.613220,377.664276 280.837189,377.448761 280.985046,377.154175 
-      z"
-      />
-      <path
-        fill="#FAF9F7"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M331.469543,95.484245 
-        C317.488678,90.165962 303.823547,85.037468 290.142700,79.951309 
-        C288.281342,79.259300 286.274963,78.937027 284.450165,78.170921 
-        C279.101257,75.925270 276.706390,71.200439 278.195282,66.139580 
-        C279.691620,61.053478 284.858215,58.137039 290.477264,59.398651 
-        C307.102539,63.131424 323.102875,68.672112 338.297577,76.412148 
-        C370.801208,92.969185 395.693665,117.612885 414.600830,148.556137 
-        C418.296478,154.604385 417.377350,160.107239 412.309784,163.440826 
-        C407.261108,166.761993 401.718750,165.460648 397.879028,159.370346 
-        C388.052460,143.784073 377.158875,128.998337 362.803833,117.372505 
-        C353.023468,109.451630 342.159668,102.868599 331.469543,95.484245 
-      z"
-      />
-      <path
-        fill="#FAF9F7"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M100.409775,368.563751 
-        C94.120689,357.445221 87.697586,346.784180 82.134781,335.691681 
-        C78.795601,329.033264 81.457748,323.319489 87.630302,320.958221 
-        C92.958626,318.919922 98.135201,321.693359 100.636574,328.434326 
-        C107.236870,346.221619 117.999825,360.907928 132.903351,372.592377 
-        C136.710892,375.577454 139.126160,379.176422 137.766647,384.222076 
-        C136.662506,388.319946 133.705093,391.110504 129.629028,391.242950 
-        C126.765190,391.336060 123.153931,390.397614 121.071205,388.563263 
-        C113.974586,382.312897 107.407326,375.461517 100.409775,368.563751 
-      z"
-      />
-      <path
-        fill="#FAF9F5"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M428.695557,186.615448 
-        C423.089508,191.441910 416.280518,191.319092 412.611084,186.633514 
-        C409.142578,182.204514 410.277100,174.733597 414.902557,171.543930 
-        C418.531738,169.041290 424.096100,169.377090 427.446198,172.300903 
-        C431.447693,175.793213 431.944855,180.461761 428.695557,186.615448 
-      z"
-      />
-      <path
-        fill="#FAF9F5"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M85.269478,297.885834 
-        C89.186470,305.516693 87.271980,311.689758 80.557129,314.246002 
-        C75.372406,316.219696 69.246780,313.054352 67.564117,307.427948 
-        C66.221504,302.938538 67.579636,299.172302 71.201172,296.341278 
-        C74.696617,293.608826 80.608101,294.235229 85.269478,297.885834 
-      z"
-      />
-      <path
-        fill="#F6F5F4"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M359.223877,315.350098 
-        C359.741608,315.105743 360.318665,315.230499 360.977417,315.649841 
-        C360.467163,315.869324 359.875183,315.794281 359.223877,315.350098 
-      z"
-      />
-      <path
-        fill="#F6F5F4"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M360.888641,316.297577 
-        C361.335114,316.090820 361.646240,316.263184 362.019409,316.705994 
-        C361.638733,316.876556 361.196045,316.776611 360.888641,316.297577 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M258.578979,380.082581 
-        C256.812866,380.590973 254.631897,381.027832 252.197906,381.172638 
-        C254.017960,380.590729 256.091034,380.300903 258.578979,380.082581 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M249.849854,380.516479 
-        C247.765564,380.875031 245.480194,380.874512 243.100555,380.513733 
-        C245.220444,380.154816 247.434616,380.156128 249.849854,380.516479 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M120.457886,242.883240 
-        C120.142166,241.750381 120.161827,240.450684 120.513916,239.078644 
-        C120.828651,240.243011 120.810959,241.479706 120.457886,242.883240 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M120.917488,259.650146 
-        C120.429779,259.102142 120.013550,258.210510 119.858109,257.132019 
-        C120.408913,257.732300 120.698936,258.519379 120.917488,259.650146 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M269.686493,379.082581 
-        C269.414978,379.551331 268.836060,379.949219 268.100586,380.112671 
-        C268.422394,379.589417 268.900757,379.300568 269.686493,379.082581 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M124.887970,283.715515 
-        C124.551300,283.837189 124.335777,283.613037 124.120193,283.089264 
-        C124.337158,283.059235 124.633484,283.252594 124.887970,283.715515 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M123.888016,279.715454 
-        C123.550880,279.836761 123.335274,279.612030 123.120621,279.086914 
-        C123.337921,279.057220 123.633911,279.251556 123.888016,279.715454 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M122.886902,274.715637 
-        C122.550911,274.837067 122.335533,274.613281 122.119659,274.090546 
-        C122.336411,274.060455 122.632530,274.253418 122.886902,274.715637 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M121.887314,268.715576 
-        C121.550728,268.836700 121.335281,268.612213 121.120300,268.087921 
-        C121.337326,268.058228 121.633232,268.252197 121.887314,268.715576 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M284.715546,376.111938 
-        C284.837189,376.448639 284.613037,376.664185 284.089172,376.879791 
-        C284.059143,376.662750 284.252533,376.366425 284.715546,376.111938 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M280.715576,377.112366 
-        C280.837189,377.448761 280.613220,377.664276 280.089935,377.880127 
-        C280.059814,377.663208 280.252930,377.366882 280.715576,377.112366 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M275.715607,378.111816 
-        C275.838196,378.448059 275.615143,378.663879 275.093323,378.880920 
-        C275.062439,378.663910 275.254242,378.367004 275.715607,378.111816 
-      z"
-      />
-      <path
-        fill="#FDFDFD"
-        opacity="1.000000"
-        stroke="none"
-        d="
-      M234.040588,379.700897 
-        C233.816666,379.889343 233.538147,379.750275 233.159653,379.346436 
-        C233.332047,379.251770 233.640823,379.348999 234.040588,379.700897 
-      z"
-      />
-    </svg>
-  );
-};
